Updated Bleeding to 1.11

+++ build/tmp/recompileMc/sources/net/minecraft/item/ItemBow.java
@@ -31,15 +31,7 @@ public class ItemBow extends Item
             @SideOnly(Side.CLIENT)
             public float apply(ItemStack stack, @Nullable World worldIn, @Nullable EntityLivingBase entityIn)
             {
-                if (entityIn == null)
-                {
-                    return 0.0F;
-                }
-                else
-                {
-                    ItemStack itemstack = entityIn.getActiveItemStack();
-                    return itemstack != null && itemstack.getItem() == Items.BOW ? (float)(stack.getMaxItemUseDuration() - entityIn.getItemInUseCount()) / 20.0F : 0.0F;
-                }
+                return entityIn == null ? 0.0F : (entityIn.getActiveItemStack().getItem() != Items.BOW ? 0.0F : (float)(stack.getMaxItemUseDuration() - entityIn.getItemInUseCount()) / 20.0F);
             }
         });
         this.addPropertyOverride(new ResourceLocation("pulling"), new IItemPropertyGetter()
@@ -74,13 +66,13 @@ public class ItemBow extends Item
                 }
             }
 
-            return null;
+            return ItemStack.field_190927_a;
         }
     }
 
-    protected boolean isArrow(@Nullable ItemStack stack)
+    protected boolean isArrow(ItemStack stack)
     {
-        return stack != null && stack.getItem() instanceof ItemArrow;
+        return stack.getItem() instanceof ItemArrow;
     }
 
     /**
@@ -98,9 +90,9 @@ public class ItemBow extends Item
             i = net.minecraftforge.event.ForgeEventFactory.onArrowLoose(stack, worldIn, (EntityPlayer)entityLiving, i, itemstack != null || flag);
             if (i < 0) return;
 
-            if (itemstack != null || flag)
+            if (!itemstack.func_190926_b() || flag)
             {
-                if (itemstack == null)
+                if (itemstack.func_190926_b())
                 {
                     itemstack = new ItemStack(Items.ARROW);
                 }
@@ -143,7 +135,7 @@ public class ItemBow extends Item
 
                         stack.damageItem(1, entityplayer);
 
-                        if (flag1)
+                        if (flag1 || entityplayer.capabilities.isCreativeMode && (itemstack.getItem() == Items.SPECTRAL_ARROW || itemstack.getItem() == Items.TIPPED_ARROW))
                         {
                             entityarrow.pickupStatus = EntityArrow.PickupStatus.CREATIVE_ONLY;
                         }
@@ -151,13 +143,13 @@ public class ItemBow extends Item
                         worldIn.spawnEntityInWorld(entityarrow);
                     }
 
-                    worldIn.playSound((EntityPlayer)null, entityplayer.posX, entityplayer.posY, entityplayer.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.NEUTRAL, 1.0F, 1.0F / (itemRand.nextFloat() * 0.4F + 1.2F) + f * 0.5F);
+                    worldIn.playSound((EntityPlayer)null, entityplayer.posX, entityplayer.posY, entityplayer.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.PLAYERS, 1.0F, 1.0F / (itemRand.nextFloat() * 0.4F + 1.2F) + f * 0.5F);
 
-                    if (!flag1)
+                    if (!flag1 && !entityplayer.capabilities.isCreativeMode)
                     {
-                        --itemstack.stackSize;
+                        itemstack.func_190918_g(1);
 
-                        if (itemstack.stackSize == 0)
+                        if (itemstack.func_190926_b())
                         {
                             entityplayer.inventory.deleteStack(itemstack);
                         }
@@ -201,21 +193,22 @@ public class ItemBow extends Item
         return EnumAction.BOW;
     }
 
-    public ActionResult<ItemStack> onItemRightClick(ItemStack itemStackIn, World worldIn, EntityPlayer playerIn, EnumHand hand)
+    public ActionResult<ItemStack> onItemRightClick(World itemStackIn, EntityPlayer worldIn, EnumHand playerIn)
     {
-        boolean flag = this.findAmmo(playerIn) != null;
+        ItemStack itemstack = worldIn.getHeldItem(playerIn);
+        boolean flag = !this.findAmmo(worldIn).func_190926_b();
 
-        ActionResult<ItemStack> ret = net.minecraftforge.event.ForgeEventFactory.onArrowNock(itemStackIn, worldIn, playerIn, hand, flag);
+        ActionResult<ItemStack> ret = net.minecraftforge.event.ForgeEventFactory.onArrowNock(itemstack, itemStackIn, worldIn, playerIn, flag);
         if (ret != null) return ret;
 
-        if (!playerIn.capabilities.isCreativeMode && !flag)
+        if (!worldIn.capabilities.isCreativeMode && !flag)
         {
-            return !flag ? new ActionResult(EnumActionResult.FAIL, itemStackIn) : new ActionResult(EnumActionResult.PASS, itemStackIn);
+            return flag ? new ActionResult(EnumActionResult.PASS, itemstack) : new ActionResult(EnumActionResult.FAIL, itemstack);
         }
         else
         {
-            playerIn.setActiveHand(hand);
-            return new ActionResult(EnumActionResult.SUCCESS, itemStackIn);
+            worldIn.setActiveHand(playerIn);
+            return new ActionResult(EnumActionResult.SUCCESS, itemstack);
         }
     }
